--- src/_includes/code/layout/listview/pubspec.yaml
@@ -1,9 +1,0 @@
-name: listview
-description: Demonstrates the ListView and ListTile widgets.
-
-dependencies:
-  flutter:
-    sdk: flutter
-
-flutter:
-  uses-material-design: true
